refactor: rename DutyHistoryRequest (#90)
- 의미를 명확하기 하기 위하여 DutyStatusUpdateRequest 로 변경
- 패키지 이동 (-> dto.request)
